quantumcue-droplet-2-api-server-current/orchestrator/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.34.0
starlette>=0.50.0
python-dotenv==1.0.1
requests>=2.31.0
scipy>=1.13.0
numpy>=1.26.0
python-multipart>=0.0.9
joblib>=1.4.0
